Move PluginLoader from __init__.py to loader.py
Move the PluginLoader class and it's global caches from
plugins/__init__.py to plugins/loader.py

One reason is so PluginLoader gets it's own module namespace
so that __name__ is more accurate so that standard log setup
will get a better logger name ('ansible.plugins.loader' vs
'ansible.plugins'). This will let logging be enabled for the
plugin loader without enabling it for all plugins.

It also makes it more clear that plugins/__init__.py is where all
the loader objects are created instead of being at the end of the
file after a large class.
